+++ lib/csv_row_model.rb
@@ -8,15 +8,16 @@ require 'active_support/all'
 require 'active_warnings'
 require 'csv'
 
-require 'csv_row_model/exceptions'
-require 'csv_row_model/inspect'
 require 'csv_row_model/validators/default_change'
 
 if autoload && defined?(Rails)
   require 'csv_row_model/engine'
 else
   require 'csv_row_model/version'
-  require 'csv_row_model/deep_class_var'
+
+  require 'csv_row_model/concerns/inspect'
+  require 'csv_row_model/concerns/deep_class_var'
+
   require 'csv_row_model/validators/validate_attributes'
 
   require 'csv_row_model/model'
@@ -33,4 +34,6 @@ else
 end
 
 module CsvRowModel
+  class RowModelClassNotDefined < StandardError; end
+  class AccessedInvalidAttribute < StandardError; end
 end
