--- lib/csv_row_model/exceptions.rb
@@ -1,4 +1,0 @@
-module CsvRowModel
-  class RowModelClassNotDefined < StandardError; end
-  class AccessedInvalidAttribute < StandardError; end
-end
