--- lib/csv_row_model/inspect.rb
@@ -1,8 +1,0 @@
-module CsvRowModel
-  module Inspect
-    def inspect
-      s = self.class.inspect_instance_variables.map { |v| "#{v}=#{instance_variable_get(v).inspect}" }.join(", ")
-      "#<#{self.class.name}:#{object_id} #{s}>"
-    end
-  end
-end
